# openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Toolbox",
        "contact": {
            "email": "strategie-navigator@jade-hs.de"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.de.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://toolbox-backend.test",
            "description": "Lokaler Entwicklungsserver"
        },
        {
            "url": "https://strategie-navigator.jade-hs.de",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/users/{user}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Zeigt einen ausgew\u00e4hlten User an",
                "operationId": "showUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Der in der Url definierte User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/complete_user"
                                                },
                                                {
                                                    "$ref": "#/components/schemas/simple_user"
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/not_found"
                    }
                }
            },
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "\u00c4ndert die Attribute eines ausgew\u00e4hlten Nutzers zu den \u00fcbergebenen",
                "operationId": "test",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Der in der Url definierte User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Nutzerdaten",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "current_password"
                                ],
                                "properties": {
                                    "username": {
                                        "description": "new username",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "new password",
                                        "type": "string"
                                    },
                                    "current_password": {
                                        "description": "current_password",
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/not_found"
                    },
                    "422": {
                        "$ref": "#/components/responses/validation_failed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "L\u00f6scht den ausgew\u00e4hlten User",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Den in der Url definierten User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/not_found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "complete_user": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/simple_user"
                    },
                    {
                        "description": "Diese Daten sind nur mit den ausreichenden Rechten verf\u00fcgbar",
                        "type": "object",
                        "properties": {
                            "email": {
                                "description": "Email des Nutzers",
                                "type": "string",
                                "nullable": true
                            },
                            "owned_saves": {
                                "description": "Id der Speicherst\u00e4nde, welche dieser Nutzer besitzt",
                                "type": "array",
                                "items": {
                                    "type": "integer"
                                }
                            },
                            "schared_saves": {
                                "description": "Id der Speicherst\u00e4nde, welchem diesem Nutzer geteilt wurden und die dieser angenommen hat",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/shared_save_of_user"
                                }
                            },
                            "invitations": {
                                "description": "Id der Speicherst\u00e4nde, zu denen dieser Nutzer eingeladen wurde und die noch nicht angenommen wurden",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/shared_save_of_user"
                                }
                            }
                        }
                    }
                ]
            },
...
            "shared_save_of_user": {
                "type": "object",
                "properties": {
                    "save_id": {
                        "description": "Id des Speicherstands, welcher geteilt wurde",
                        "type": "integer"
                    },
                    "permission": {
                        "description": "Rechte die die geteilte person besitzt. 0: Leserechte \\t 1: Schreib- und Leserechte \\t 2: Adminrechte",
                        "enum": [
                            0,
                            1,
                            2
                        ],
                        "type": "integer"
                    }
                }
            },
            "simple_user": {
                "description": "Dieses Schema wird zur\u00fcck gegeben, wenn der aktuelle Nutzer nicht genugRechte hat um private informationen \u00fcber den Nutzer zu sehen",
                "type": "object",
                "properties": {
                    "id": {
                        "description": "Id des Nutzers",
                        "type": "integer",
                        "example": 1
                    },
                    "username": {
                        "description": "Name des nutzers",
                        "type": "string",
                        "example": "max musterman"
                    },
                    "anonymous": {
                        "description": "Gibt an ab dieser User ein anonymer nutzer ist",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Zeitpunkt an dem dieses Nutzerkonto erstellt wurde. Wenn dieses Nutzerkonto aus einem anonymen Konto erstellt wurde, wird das erstelldatum den anonymen Kontos angezeigt",
                        "format": "date-time",
                        "type": "string"
                    }
                }
            }
        },
        "responses": {
...
            "not_found": {
                "description": "Entity not found"
            },
            "unauthenticated": {
                "description": "The current user isn't authenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "default": "Unauthenticated",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "unauthorized": {
                "description": "The current user isn't authorized to do this action",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "default": "This action is unauthorized.",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "validation_failed": {
                "description": "Ein oder mehrere Argumente sind invalide",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "description": "Eine Nachricht die einem Nutzer angezeigt werden k\u00f6nnte",
                                    "type": "string"
                                },
                                "errors": {
                                    "description": "Keys des Objects haben den selben namen wie in dem Request Body",
                                    "type": "array",
                                    "items": {
                                        "description": "Fehlermeldungen eines Arguments. Der Key hat den selben namen wie das Argument",
                                        "type": "object",
                                        "additionalProperties": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "BearerToken": []
        }
    ],
...
}